fixed error with computer player input parse
the generateWord method should work correctly with everything else now. it ended up being an issue with the way the input files are formatted vs the dictionary's format. it looks like there are carriage returns in the dictionary, which was causing an incorrect sum when strcmp was used. segfault issue on 5 used words still exists